demo-usecase-2-AnomalyPrediction-Python-SDK/Predictor/requirements.txt
asn1crypto==0.24.0
azure-common==1.1.16
azure-functions==1.0.0a5
azure-functions-worker==1.0.0a6
azure-storage-blob==1.4.0
azure-storage-common==1.4.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
grpcio==1.14.2
grpcio-tools==1.14.2
idna==2.8
numpy==1.15.4
protobuf==3.6.1
pycparser==2.19
python-dateutil==2.7.5
requests==2.21.0
scikit-learn==0.19.1
scipy==1.2.0
six==1.12.0
sklearn==0.0
urllib3==1.24.1

